Create generator-generic-ossf-slsa3-publish.yml
An aviator is a skilled professional who pilots or navigates aircraft, such as airplanes or helicopters. They are trained to ensure safe and efficient air travel, often symbolizing adventure, innovation, and mastery of flight. The term also extends to military pilots and cultural icons, as well as the classic aviator sunglasses.